+++ fluentlenium-core/src/test/java/org/fluentlenium/integration/shareddriver/SharedDriverDeleteCookies.java
@@ -19,25 +19,31 @@ import org.fluentlenium.integration.localtest.LocalFluentCase;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
+import org.openqa.selenium.Cookie;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.fluentlenium.core.filter.FilterConstructor.withName;
 
-@SharedDriver(type = SharedDriver.SharedType.PER_CLASS)
+@SharedDriver(deleteCookies = true)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class SharedDriverPerClass1Test extends LocalFluentCase {
+public class SharedDriverDeleteCookies extends LocalFluentCase {
+
 
   @Test
-  public void firstMethod() {
+  public void cookieFirstMethod() {
     goTo(LocalFluentCase.DEFAULT_URL);
     assertThat($(".small", withName("name"))).hasSize(1);
+    this.getDriver().manage().addCookie(new Cookie("cookie", "fluent"));
   }
 
 
   @Test
-  public void secondMethod() {
+  public void cookieSecondMethod() {
     assertThat($(".small", withName("name"))).hasSize(1);
+    assertThat(this.getCookie("cookie")).isNull();
   }
+
+
 
 
 }
