+++ fluentlenium-core/src/test/java/org/fluentlenium/integration/shareddriver/SharedDriverOnce2.java
@@ -16,29 +16,22 @@ package org.fluentlenium.integration.shareddriver;
 
 import org.fluentlenium.adapter.util.SharedDriver;
 import org.fluentlenium.integration.localtest.LocalFluentCase;
-import org.junit.FixMethodOrder;
 import org.junit.Test;
-import org.junit.runners.MethodSorters;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.fluentlenium.core.filter.FilterConstructor.withName;
 
-@SharedDriver(type = SharedDriver.SharedType.PER_METHOD)
-@FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class SharedDriverPerMethodByAnnotationTest extends LocalFluentCase {
+@SharedDriver
+public class SharedDriverOnce2 extends LocalFluentCase {
 
 
-  @Test
-  public void firstMethod() {
-    goTo(LocalFluentCase.DEFAULT_URL);
-    assertThat($(".small", withName("name"))).hasSize(1);
-  }
 
 
   @Test
   public void secondMethod() {
-    assertThat($(".small", withName("name"))).hasSize(0);
+    assertThat($(".small", withName("name"))).hasSize(1);
   }
+
