+++ fluentlenium-core/src/test/java/org/fluentlenium/integration/shareddriver/SharedDriverPerClass2.java
@@ -14,30 +14,22 @@
 
 package org.fluentlenium.integration.shareddriver;
 
+import org.fluentlenium.adapter.util.SharedDriver;
 import org.fluentlenium.integration.localtest.LocalFluentCase;
-import org.junit.FixMethodOrder;
 import org.junit.Test;
-import org.junit.runners.MethodSorters;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.fluentlenium.core.filter.FilterConstructor.withName;
 
-@FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class SharedDriverPerMethodByDefaultTest extends LocalFluentCase {
+@SharedDriver(type = SharedDriver.SharedType.PER_CLASS)
+public class SharedDriverPerClass2 extends LocalFluentCase {
 
-
-  @Test
-  public void firstMethod() {
-    goTo(LocalFluentCase.DEFAULT_URL);
-    assertThat($(".small", withName("name"))).hasSize(1);
-  }
 
 
   @Test
   public void secondMethod() {
     assertThat($(".small", withName("name"))).hasSize(0);
   }
-
 
 
 }
