+++ fluentlenium-core/src/test/java/org/fluentlenium/integration/shareddriver/SharedDriverPerMethodByDefault.java
@@ -14,35 +14,29 @@
 
 package org.fluentlenium.integration.shareddriver;
 
-import org.fluentlenium.adapter.util.SharedDriver;
 import org.fluentlenium.integration.localtest.LocalFluentCase;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
-import org.openqa.selenium.Cookie;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.fluentlenium.core.filter.FilterConstructor.withName;
 
-@SharedDriver(deleteCookies = true)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class SharedDriverDeleteCookiesTest extends LocalFluentCase {
+public class SharedDriverPerMethodByDefault extends LocalFluentCase {
 
 
   @Test
-  public void cookieFirstMethod() {
+  public void firstMethod() {
     goTo(LocalFluentCase.DEFAULT_URL);
     assertThat($(".small", withName("name"))).hasSize(1);
-    this.getDriver().manage().addCookie(new Cookie("cookie", "fluent"));
   }
 
 
   @Test
-  public void cookieSecondMethod() {
-    assertThat($(".small", withName("name"))).hasSize(1);
-    assertThat(this.getCookie("cookie")).isNull();
+  public void secondMethod() {
+    assertThat($(".small", withName("name"))).hasSize(0);
   }
-
