+++ fluentlenium-core/src/test/java/org/fluentlenium/integration/shareddriver/SharedDriverSuiteTest.java
@@ -14,22 +14,18 @@
 
 package org.fluentlenium.integration.shareddriver;
 
-import org.fluentlenium.adapter.util.SharedDriver;
-import org.fluentlenium.integration.localtest.LocalFluentCase;
-import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.fluentlenium.core.filter.FilterConstructor.withName;
-
-@SharedDriver(type = SharedDriver.SharedType.PER_CLASS)
-public class SharedDriverPerClass2Test extends LocalFluentCase {
-
-
-
-  @Test
-  public void secondMethod() {
-    assertThat($(".small", withName("name"))).hasSize(0);
-  }
-
-
+@RunWith(Suite.class)
+@Suite.SuiteClasses(value = {
+    SharedDriverDeleteCookies.class,
+    SharedDriverOnce1.class,
+    SharedDriverOnce2.class,
+    SharedDriverPerClass1.class,
+    SharedDriverPerClass2.class,
+    SharedDriverPerMethodByAnnotation.class,
+    SharedDriverPerMethodByDefault.class
+})
+public class SharedDriverSuiteTest {
 }
